cli unit tests
