+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "cargo-hoist"
 description = "Dead simple, memoized cargo subcommand to hoist cargo-built binaries into scope."
-version = "0.1.6"
+version = "0.1.7"
 edition = "2021"
 license = "MIT"
 authors = ["refcell"]
@@ -37,3 +37,5 @@ termcolor = "1.3.0"
 [dev-dependencies]
 tempfile = "3.8"
 serial_test = "2.0"
+assert_cmd = "2.0.12"
+rand = "0.8.5"
